requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
python-dotenv==1.0.1
faiss-cpu==1.7.4.post2
numpy==1.26.4
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-docx==1.1.2
groq==0.11.0
google-generativeai==0.7.2
